+++ backend/app/models/transference.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 import uuid
 from sqlmodel import Field, Relationship
 
@@ -7,7 +8,7 @@ from app.db import SQLBaseModel, SQLBaseModelAudit
 class Transference(SQLBaseModelAudit, table=True):
     id: int = Field(default=None, primary_key=True)
     number: uuid.UUID = Field(default_factory=uuid.uuid4, unique=True, nullable=False)
-    balance: float = Field(default=0.0, ge=0.0, nullable=False)
+    balance: Decimal = Field(default=0.00, ge=0.00, nullable=False, max_digits=15, decimal_places=2)
     from_account_number: uuid.UUID = Field(foreign_key="account.number", nullable=False)
     to_account_number: uuid.UUID = Field(foreign_key="account.number", nullable=False)
 
@@ -24,12 +25,12 @@ class Transference(SQLBaseModelAudit, table=True):
 class TransferenceRead(SQLBaseModel):
     id: int = Field(default=None, primary_key=True)
     number: uuid.UUID = Field(default_factory=uuid.uuid4, unique=True, nullable=False)
-    balance: float = Field(default=0.0, ge=0.0, nullable=False)
+    balance: Decimal = Field(default=0.00, ge=0.00, nullable=False, max_digits=15, decimal_places=2)
     from_account_number: uuid.UUID = Field(foreign_key="account.number", nullable=False)
     to_account_number: uuid.UUID = Field(foreign_key="account.number", nullable=False)
 
 
 class TransferenceCreate(SQLBaseModel):
-    balance: float = Field(default=0.0, ge=0.0, nullable=False)
+    balance: Decimal = Field(default=0.00, ge=0.00, nullable=False, max_digits=15, decimal_places=2)
     from_account_number: uuid.UUID = Field(foreign_key="account.number", nullable=False)
     to_account_number: uuid.UUID = Field(foreign_key="account.number", nullable=False)
